use absolute paths and refactor
