+++ src/MiniJuvix/Prelude/Error/GenericError.hs
@@ -25,10 +25,15 @@ instance Pretty GenericError where
   pretty :: GenericError -> Doc a
   pretty g =
     let lineNum = g ^. genericErrorLoc . locFileLoc . locLine
-        colNum = g ^. genericErrorLoc . locFileLoc . locCol in
-    pretty (g ^. genericErrorFile)
-    <> colon <> pretty lineNum <> colon <> pretty colNum
-    <> colon <+> "error" <> colon <+> pretty (g ^. genericErrorMessage) <> line
+        colNum = g ^. genericErrorLoc . locFileLoc . locCol
+     in pretty (g ^. genericErrorFile)
+          <> colon
+          <> pretty lineNum
+          <> colon
+          <> pretty colNum
+          <> colon <+> "error"
+          <> colon <+> pretty (g ^. genericErrorMessage)
+          <> line
 
 renderGenericError :: GenericError -> Text
 renderGenericError = renderStrict . layoutPretty defaultLayoutOptions . pretty
